refactor(package): 패키징 변경
