+++ app/src/main/java/woowacourse/app/model/advertisement/AdvertisementUiModel.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import android.net.Uri
 import androidx.annotation.DrawableRes
 import woowacourse.app.model.MainData
-import woowacourse.app.ui.main.adapter.MainViewType
+import woowacourse.app.ui.main.home.adapter.MainViewType
 
 data class AdvertisementUiModel(
     override val id: Long,
