+++ app/src/main/java/woowacourse/app/ui/main/home/adapter/listview/AdvertisementViewHolder.kt
@@ -1,11 +1,11 @@
-package woowacourse.app.ui.main.adapter.recyclerview
+package woowacourse.app.ui.main.home.adapter.listview
 
 import android.content.Intent
 import android.view.View
 import android.widget.ImageView
 import woowacourse.app.model.MainData
 import woowacourse.app.model.advertisement.AdvertisementUiModel
-import woowacourse.app.ui.main.adapter.MainViewType
+import woowacourse.app.ui.main.home.adapter.MainViewType
 import woowacourse.movie.R
 
 class AdvertisementViewHolder(view: View) : MainViewHolder(view) {
@@ -19,7 +19,7 @@ class AdvertisementViewHolder(view: View) : MainViewHolder(view) {
         poster.setImageResource(advertisement.image)
     }
 
-    fun setAdvertisementClick(clickAd: (Intent) -> Unit) {
+    fun clickAdvertisement(clickAd: (Intent) -> Unit) {
         view.setOnClickListener {
             clickAd(advertisement.getIntent())
         }
