+++ app/src/main/java/woowacourse/app/ui/main/home/adapter/listview/MainViewHolder.kt
@@ -1,8 +1,8 @@
-package woowacourse.app.ui.main.adapter.listview
+package woowacourse.app.ui.main.home.adapter.listview
 
 import android.view.View
 import woowacourse.app.model.MainData
-import woowacourse.app.ui.main.adapter.MainViewType
+import woowacourse.app.ui.main.home.adapter.MainViewType
 
 sealed class MainViewHolder(val view: View) {
     abstract fun onBind(data: MainData)
