+++ app/src/main/java/woowacourse/app/ui/main/home/adapter/recyclerview/AdvertisementViewHolder.kt
@@ -1,14 +1,14 @@
-package woowacourse.app.ui.main.adapter.listview
+package woowacourse.app.ui.main.home.adapter.recyclerview
 
 import android.content.Intent
 import android.view.View
 import android.widget.ImageView
 import woowacourse.app.model.MainData
 import woowacourse.app.model.advertisement.AdvertisementUiModel
-import woowacourse.app.ui.main.adapter.MainViewType
+import woowacourse.app.ui.main.home.adapter.MainViewType
 import woowacourse.movie.R
 
-class AdvertisementViewHolder(view: View) : MainViewHolder(view) {
+class AdvertisementViewHolder(view: View) : HomeViewHolder(view) {
     private val poster: ImageView = view.findViewById(R.id.imageAdvertisement)
 
     override val mainViewType: MainViewType = MainViewType.ADVERTISEMENT
@@ -19,7 +19,7 @@ class AdvertisementViewHolder(view: View) : MainViewHolder(view) {
         poster.setImageResource(advertisement.image)
     }
 
-    fun clickAdvertisement(clickAd: (Intent) -> Unit) {
+    fun setAdvertisementClick(clickAd: (Intent) -> Unit) {
         view.setOnClickListener {
             clickAd(advertisement.getIntent())
         }
