+++ app/src/main/java/woowacourse/app/ui/main/home/adapter/recyclerview/HomeAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.app.ui.main.adapter.recyclerview
+package woowacourse.app.ui.main.home.adapter.recyclerview
 
 import android.content.Context
 import android.content.Intent
@@ -6,20 +6,20 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.app.model.MainData
-import woowacourse.app.ui.main.adapter.MainViewType
+import woowacourse.app.ui.main.home.adapter.MainViewType
 import woowacourse.movie.R
 
-class MainAdapter(
+class HomeAdapter(
     context: Context,
     private val clickBook: (Long) -> Unit,
     private val clickAd: (Intent) -> Unit,
-) : RecyclerView.Adapter<MainViewHolder>() {
+) : RecyclerView.Adapter<HomeViewHolder>() {
     private val movies = mutableListOf<MainData>()
     private val layoutInflater = LayoutInflater.from(context)
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeViewHolder {
         val mainViewType = MainViewType.getMainViewType(viewType)
-        val viewHolder: MainViewHolder = when (mainViewType) {
+        val viewHolder: HomeViewHolder = when (mainViewType) {
             MainViewType.CONTENT -> {
                 MovieViewHolder(layoutInflater.inflate(R.layout.movie_list_item, parent, false))
             }
@@ -33,14 +33,14 @@ class MainAdapter(
         return viewHolder
     }
 
-    private fun setViewHolderClick(viewHolder: MainViewHolder) {
+    private fun setViewHolderClick(viewHolder: HomeViewHolder) {
         when (viewHolder) {
             is MovieViewHolder -> viewHolder.setBookingClick(clickBook)
             is AdvertisementViewHolder -> viewHolder.setAdvertisementClick(clickAd)
         }
     }
 
-    override fun onBindViewHolder(holder: MainViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: HomeViewHolder, position: Int) {
         holder.onBind(movies[position])
     }
