+++ app/src/main/java/woowacourse/app/ui/main/home/adapter/recyclerview/HomeViewHolder.kt
@@ -1,11 +1,11 @@
-package woowacourse.app.ui.main.adapter.recyclerview
+package woowacourse.app.ui.main.home.adapter.recyclerview
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.app.model.MainData
-import woowacourse.app.ui.main.adapter.MainViewType
+import woowacourse.app.ui.main.home.adapter.MainViewType
 
-sealed class MainViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
+sealed class HomeViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
     abstract fun onBind(data: MainData)
     abstract val mainViewType: MainViewType
 }
