+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="woowacourse.app.ui.main.HomeFragment">
+    tools:context="woowacourse.app.ui.main.home.HomeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/listMainMovie"
